requirements.txt
# --- Core Web + Parsing ---
requests>=2.31.0
beautifulsoup4>=4.12.0

# --- Image Processing ---
Pillow>=10.0.0

# --- Browser Automation ---
selenium>=4.15.0
webdriver-manager>=4.0.0

# --- Database ---
psycopg2-binary>=2.9.9

# --- Google Cloud Clients ---
google-cloud-storage>=2.15.0
google-auth>=2.23.0
google-auth-oauthlib>=1.2.0

# --- Utilities ---
tqdm>=4.66.0
